// frontend/src/app/core/api/model/userCreationRequestDto.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface UserCreationRequestDto { 
    email: string;
    name: string;
    password: string;
    role: UserCreationRequestDtoRoleEnum;
    teamId: number;
    department: UserCreationRequestDtoDepartmentEnum;
    promoteToTeamLeader?: boolean;
}
export enum UserCreationRequestDtoRoleEnum {
    Requester = 'REQUESTER',
    ProcurementOfficer = 'PROCUREMENT_OFFICER',
    FinanceOfficer = 'FINANCE_OFFICER',
    Administrator = 'ADMINISTRATOR'
};
export enum UserCreationRequestDtoDepartmentEnum {
    It = 'IT',
    Rd = 'RD',
    Hr = 'HR',
    Sales = 'SALES',
    Legal = 'LEGAL'
};