// frontend/src/app/core/api/model/teamDto.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { UserDto } from './userDto';


export interface TeamDto { 
    id?: number;
    name?: string;
    description?: string;
    department?: string;
    leaderId?: number;
    members?: Array<UserDto>;
    isActive?: boolean;
    expiresAt?: string;
    createdAt?: string;
}